+++ pages/dashboard/components/storageHistoryChart.vue
@@ -8,6 +8,7 @@
 <script setup>
 import { onMounted, onBeforeMount } from 'vue'
 import { Chart } from '@antv/g2';
+import { paseDate } from '../common'
 
 const storageHistoryData = ref([])
 
@@ -42,23 +43,6 @@ const getStorageHistoryData = async () => {
   })
 }
 
-
-const paseDate = (val) => {
-  val = val * 1
-  let time = new Date(val)
-  let y = time.getFullYear();
-  let m = time.getMonth() + 1
-  let d = time.getDate()
-  let h = time.getHours()
-  let mm = time.getMinutes()
-  let s = time.getSeconds()
-  return `${y}-${checkAddZero(m)}-${checkAddZero(d)} ${checkAddZero(h)}:${checkAddZero(mm)}:${checkAddZero(s)}`
-}
-
-const checkAddZero = (m) => {
-  return m < 10 ? '0' + m : m
-}
-
 const initChart = () => {
   const chart = new Chart({
     container: 'storag-history-chart',
@@ -88,12 +72,22 @@ const initChart = () => {
     .area()
     .adjust('stack')
     .position('time*value')
-    .color('item', ['l(100) 0:#FF8A00 1:#ffffff', 'l(100) 0:#57D2B4 1:#ffffff', 'l(100) 0:#B835F5 1:#ffffff']);
+    .color('item', ['l(90) 0:rgba(255, 153, 0, 0.45), 1:rgba(46, 42, 40, 0)',
+      'l(90) 0:rgba(0, 255, 163, 0.31) 1:rgba(46, 42, 40, 0)',
+      'l(90) 0:rgba(155, 38, 203, 1) 1:rgba(46, 42, 40, 0.53)'])
+    .tooltip('item*value', (item, value) => {
+      value = value + 'TB';
+      return {
+        name: item,
+        value: value,
+      };
+    });
   chart
     .line()
     .adjust('stack')
     .position('time*value')
-    .color('item', ['#FF8A00', '#57D2B4', '#B835F5']);
+    .color('item', ['#FF8A00', '#57D2B4', '#B835F5'])
+    .tooltip(false);
 
   chart.interaction('element-highlight');
