+++ pages/dashboard/index.vue
@@ -1,8 +1,8 @@
 <template>
   <div class="main-margin sm:mt-[160px] mt-[120px]">
     <div>
-      <div class="text-[40px] mb-[16px] text-center font-bold">HAMSTER Dashboard</div>
-      <div class="area-desc">Explore the latest stats on the HAMSTER network via the HAMSTER dashboard
+      <div class="text-[40px] mb-[4px] text-center font-bold">HAMSTER Dashboard</div>
+      <div class="area-desc text-[16px]">Explore the latest stats on the HAMSTER network via the HAMSTER dashboard
       </div>
     </div>
 
@@ -15,7 +15,7 @@
       </div>
     </div>
 
-    <div class="mt-[64px] grid sm:grid-cols-1 md:grid-cols-3 xl:gap-[24px] gap-[8px]">
+    <div class="mt-[24px] grid sm:grid-cols-1 md:grid-cols-3 xl:gap-[24px] gap-[8px]">
       <div class="flex flex-col sm:col-span-1 md:col-span-1 bg-[#2E2A28] rounded-[16px] h-[418px] sm:h-auto p-[30px]">
         <BarChart />
       </div>
@@ -24,7 +24,7 @@
       </div>
     </div>
 
-    <div class=" mt-[64px] grid sm:grid-cols-1 md:grid-cols-3 xl:gap-[24px] gap-[8px]">
+    <div class=" mt-[24px] grid sm:grid-cols-1 md:grid-cols-3 xl:gap-[24px] gap-[8px]">
       <div class="flex flex-col sm:col-span-1 md:col-span-1 bg-[#2E2A28] rounded-[16px] h-[418px] sm:h-auto p-[30px]">
         <RAMBarChart />
       </div>
@@ -33,7 +33,7 @@
       </div>
     </div>
 
-    <div class="mt-[64px] grid sm:grid-cols-1 md:grid-cols-3 xl:gap-[24px] gap-[8px]">
+    <div class="mt-[24px] grid sm:grid-cols-1 md:grid-cols-3 xl:gap-[24px] gap-[8px]">
       <div class="flex flex-col sm:col-span-1 md:col-span-1 bg-[#2E2A28] rounded-[16px] h-[418px] sm:h-auto p-[30px]">
         <StorageBarChart />
       </div>
@@ -44,7 +44,6 @@
   </div>
 </template>
 <script setup>
-import { ref, watch, onBeforeMount, onMounted } from 'vue'
 import BinChart from './components/binChart.vue'
 import BarChart from './components/barChart.vue'
 import AreaChart from './components/areaChart.vue'
@@ -59,24 +58,6 @@ definePageMeta({
   layout: "no-ssr"
 })
 
-const flag = ref(false)
-console.log(flag.value)
-
-onBeforeMount(() => {
-
-})
-
-// const ff = () => {
-//   const url = '/hamster/dashboard/cpu-history'
-
-//   $fetch(url, {
-//     method: "GET",
-//   }).then((res) => {
-//     console.log(res)
-//   }).catch((err) => {
-//     // console.log(err)
-//   })
-// }
 </script>
 
 <style scoped>
